+++ lib/gen/dartling_test.dart
@@ -273,13 +273,9 @@ Attribute findIdAttribute(Concept concept) {
 
 String setAttributeRandomly(Attribute attribute, String entity) {
   var sc = '';
-  if (attribute.type.code == 'String') {
-    if (attribute.type.origin == 'Email') {
-      sc = '${sc}  ${entity}.${attribute.code} = ${randomEmail()}; \n';      
-    } else {
-      sc = '${sc}  ${entity}.${attribute.code} = ${randomWord()}; \n';
-    }
-  } else if (attribute.type.code == 'num') {
+  if (attribute.type.code == 'String') {     
+    sc = '${sc}    ${entity}.${attribute.code} = "${randomWord()}"; \n';      
+  }  else if (attribute.type.code == 'num') {
     sc = '${sc}  ${entity}.${attribute.code} = ${randomNum(1000)}; \n';
   } else if (attribute.type.code == 'int') {
     sc = '${sc}  ${entity}.${attribute.code} = ${randomInt(10000)}; \n';
@@ -290,8 +286,9 @@ String setAttributeRandomly(Attribute attribute, String entity) {
   } else if (attribute.type.code == 'DateTime') {
     sc = '${sc}  ${entity}.${attribute.code} = new DateTime.now(); \n';
   } else if (attribute.type.code == 'Uri') {
-    sc = '${sc}  ${entity}.'
-         '${attribute.code} = Uri.parse("${randomUri()}"); \n';
+    sc = '${sc}  ${entity}.${attribute.code} = Uri.parse("${randomUri()}"); \n';
+  } else if (attribute.type.code == 'Email') {
+    sc = '${sc}  ${entity}.${attribute.code} = ${randomEmail()}); \n';
   } else {
     sc = '${sc}  ${entity}.${attribute.code} = ${randomWord()}; \n';
   }