+++ CHANGELOG.md
@@ -4,6 +4,9 @@ based on [Semantic Versioning](http://semver.org/)
 
 *2.0.0* 2014-03-09
 
++ for Email type generate random init values in model
++ support type code vs. type base (Email vs. String)
++ decomment Email and Other types in domain
 + concept external parents and external required parents
 + model ordered (by external parent count) entry concepts
 + generate model init with initChildren ordered by external parent count